requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
sqlalchemy
jinja2